test(devtools): assert hook runtime parity via behavior, not byte equality
The previous drift test imported the built dist codegen artifact, which is not
present when unit tests run in CI, and it compared regenerated output byte for
byte. That output depends on the bundler: Function.prototype.toString() differs
between tsdown (minified) and esbuild (vitest/tsx), so equality is unstable.

Replace it with a single parametrized suite that runs one behavioral contract
against both the canonical installer and the committed browser-extension
devtools-hook.js (executed in a stubbed window). A stale or hand-edited file
fails the contract, and there is no dependency on built artifacts.
